Update jsConfig and support filepaths
with netlify functions there seems to be a problem with the
import path and how we import stuffs. In order to get the whole dev
team on the same page, I created a settings.json that specifies how the
import should work for BE. If relative imports are not used, we
should be fine.