Updated ARPS to handle email properly
After a long search, we found out that the calculations that we were making were
not as we had thought it out to be. With so many variations, calculating rent itself
became very complex. We wanted to trim that excess fat down, make it readable and
hopefully make it understandable.