Added the ability to add apiFields into the template
- Added util function to build apiFields into the provided template
- Added support for signer fields and also added some cleanup.

Added support to view checkboxes in correct format

- after the initial review, we found out the checkboxes on the goodsign are expected to contain the tick, we
were passing true or false and it was just printing true or false out. rather it should be tick or nothing.

- also we needed to clean up the esign template since each field is unique. hence we have to pass fields as
how the page is expected to be drafted. I thought we could pass dynamic values but that does not seem to be the case.
Added support for this revamp.